pyproject.toml
[tool.poetry]
name = "neuro-key"
version = "0.1.0"
description = "Unsupervised neural keyword extraction"
authors = ["Shahrukh Khan <sk28671@gmail.com>"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.10"
hydra-core = "^1.3.2"
nltk = "^3.8.1"
pydantic = "^2.3.0"
sentence-transformers = "^2.2.2"
spacy = "^3.6.1"
loguru = "^0.7.2"
srsly = "^2.4.7"

[tool.black]
line-length = 130
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
)/
'''

[tool.flake8]
max-line-length = 130
extend-ignore = ["D203", "E203", "E251", "E266", "E302", "E305", "E401", "E402", "E501", "F401", "F403", "W503"]
exclude = [".git", "__pycache__", "dist"]
max-complexity = 10

[tool.isort]
atomic = true
profile = "black"
line_length = 130
skip_gitignore = true
